+++ src/main/main.c
@@ -6,51 +6,13 @@
 /*   By: bchedru <bchedru@student.42lehavre.fr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/30 15:09:02 by tom               #+#    #+#             */
-/*   Updated: 2024/11/07 17:37:02 by bchedru          ###   ########.fr       */
+/*   Updated: 2024/11/12 16:54:20 by bchedru          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
 
 volatile int	g_exit_code = 0;
-
-t_env	*init_env(char **envp)
-{
-	t_env	*env;
-	int		i;
-
-	i = -1;
-	env = ft_calloc(1, sizeof(t_env));
-	env->envv = ft_calloc(double_array_size(envp) + 1, sizeof(char *));
-	env->quote = false;
-	while (envp[++i])
-	{
-		env->envv[i] = ft_strdup(envp[i]);
-		if (errno == ENOMEM)
-		{
-			ft_free_double_array(env->envv);
-			return (NULL);
-		}
-	}
-	i = -1;
-	while (env->envv[++i])
-	{
-		if (ft_strncmp(env->envv[i], "PWD", 3) == 0)
-		{
-			env->pwd = ft_strdup(env->envv[i]);
-			env->pwd_position = i;
-		}
-		if (ft_strncmp(env->envv[i], "OLDPWD", 6) == 0)
-		{
-			env->oldpwd = ft_strdup(env->envv[i]);
-			env->oldpwd_position = i;
-		}
-		if (ft_strncmp(env->envv[i], "HOME", 4) == 0)
-			env->home_position = i;
-	}
-	return (env);
-}
-
 
 void	sigint_handler(int signal)
 {
@@ -63,27 +25,10 @@ void	sigint_handler(int signal)
 	}
 }
 
-bool	loop(t_env *env)
+static void	loop_bis(char *line, t_ast *ast, t_env *env)
 {
-	char	*line;
-	t_ast	*ast;
 	char	**temp;
 
-	line = readline("minicheh-> ");
-	if (!line)
-		ft_exit(NULL, NULL, env, NULL);
-	if (line[0] == '\0')
-		return (false);
-	add_history(line);
-	ast = ft_calloc(1, sizeof(t_ast) + 1);
-	ast->base = ft_calloc(1, sizeof(t_ast_content) + 1);
-	env->nb_commands = 0;
-	if (only_wspace(line))
-	{
-		free(line);
-		free_ast(ast);
-		return (false);
-	}
 	parse(line, &ast, env, 0);
 	free(line);
 	line = NULL;
@@ -97,6 +42,32 @@ bool	loop(t_env *env)
 	ft_export(temp, &env);
 	ft_free_double_array(temp);
 	rl_replace_line("", 0);
+}
+
+bool	loop(t_env *env)
+{
+	char	*line;
+	t_ast	*ast;
+
+	line = readline("minicheh-> ");
+	if (!line)
+	{
+		ft_putstr_fd("exit\n", STDERR_FILENO);
+		ft_exit(NULL, NULL, env, NULL);
+	}
+	if (line[0] == '\0')
+		return (false);
+	add_history(line);
+	ast = ft_calloc(1, sizeof(t_ast) + 1);
+	ast->base = ft_calloc(1, sizeof(t_ast_content) + 1);
+	env->nb_commands = 0;
+	if (only_wspace(line))
+	{
+		free(line);
+		free_ast(ast);
+		return (false);
+	}
+	loop_bis(line, ast, env);
 	return (true);
 }
 
@@ -119,7 +90,6 @@ int	main(int ac, char **av, char **envp)
 		errno = 0;
 		signal(SIGINT, &sigint_handler);
 		signal(SIGQUIT, SIG_IGN);
-		signal(SIGPIPE, SIG_IGN);
 		if (!loop(env))
 			continue ;
 	}
